backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.0
requests>=2.31.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
langchain-core>=0.1.0

# RAG 相关
faiss-cpu>=1.7.4
tiktoken>=0.5.0
pymilvus>=2.3.0
